add: コメントの削除を管理者(user_id=1)若しくは作成者のみ行えるようにした.

+++ app/Http/Controllers/CommentController.php
@@ -113,8 +113,13 @@ class CommentController extends Controller
      */
     public function delete(Comment $comment)
     {
-        $comment->delete();
-        return redirect('/snacks/' . $comment->snack->id);
+        $user = auth()->user();
+        if($user->can('update', $comment)){
+            $comment->delete();
+            return redirect('/snacks/' . $comment->snack->id);
+        } else {
+            return redirect('/comments/' . $comment->id);
+        }
     }
     
     public function bookmarked()
